requirements.txt

pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
minio
dbt-core
dbt-postgres
s3fs
aiobotocore
boto3
pyarrow
fastparquet
psycopg2-binary
pyspark
findspark
pyspark==3.4.1
delta-spark==2.4.0
